Fixed "Path is Outside The Bas Path Issue"

+++ chat/views.py
@@ -1,10 +1,11 @@
+from chat.utils import save_temp_profile_image_from_base64String
+from loginmodule.forms import AccountUpdateForm
 from friends.views import friend_list_view
 from friends.models import FriendList, FriendRequest
 from django.shortcuts import redirect, render
-from loginmodule.models import User
+from loginmodule.models import User, get_default_profile_image
 from django.http import HttpResponse
 from django.conf import settings
-from django.core.files.storage import FileSystemStorage
 from django.core import files
 from friends.utils import get_friend_request_or_false
 from friends.friends_request_status import FriendRequestStatus
@@ -12,9 +13,6 @@ from friends.models import FriendList
 import os
 import cv2
 import json
-import base64
-TEMP_PROFILE_IMAGE_NAME = "temp_profile_image.png"
-
 # Create your views here.
 def home(request):
     return render(request,"chat/home.html",{"fullname": request.user.username})
@@ -124,13 +122,13 @@ def edit_account_view(request,*args,**kwargs):
     context = {}
     context['DATA_UPLOAD_MAX_MEMORY_SIZE'] = settings.DATA_UPLOAD_MAX_MEMORY_SIZE
     if request.POST:
-        form = userUpdateForm(request.POST,request.FILES,instance=request.user)
+        form = AccountUpdateForm(request.POST,request.FILES,instance=request.user)
 
         if form.is_valid():
             form.save()
             return redirect("account:view",user_id=user.pk)
         else:
-            form = userUpdateForm(request.POST, instance=request.user,
+            form = AccountUpdateForm(request.POST, instance=request.user,
             initial= {
                 "id": user.pk,
                 "email": user.email,
@@ -141,7 +139,7 @@ def edit_account_view(request,*args,**kwargs):
             )
             context['form'] = form
     else:
-        form = userUpdateForm(
+        form = AccountUpdateForm(
             initial= {
                 "id": user.pk,
                 "email": user.email,
@@ -153,32 +151,9 @@ def edit_account_view(request,*args,**kwargs):
         context['form'] = form
 
     
-    return render(request,"account/edit_user.html", context)
+    return render(request,"account/edit_account.html", context)
     
 
-
-def save_temp_profile_image_from_base64String(imageString, user):
-    INCORRECT_PADDING_EXCEPTION = "Incorrect padding"
-
-    try:
-
-        if not os.path.exists(settings.TEMP):
-            os.mkdir(settings.TEMP)
-        if not os.path.exists(f"{settings.TEMP}/{str(user.pk)}"):
-            os.mkdir(settings.TEMP + "/" + str(user.pk))
-        url = os.path.join(settings.TEMP + "/" + str(user.pk),TEMP_PROFILE_IMAGE_NAME)
-        storage = FileSystemStorage(location=url)
-        image = base64.b64decode(imageString)
-        with storage.open('', 'wb+') as destination:
-            destination.write(image)
-            destination.close()
-
-        return url
-    except Exception as e:
-        if str(e) == INCORRECT_PADDING_EXCEPTION:
-            imageString += "=" * ((4 - len(imageString) % 4) % 4)
-            return save_temp_profile_image_from_base64String(imageString,user)
-        return None
 
 def crop_image(request,*args,**kwargs):
     payload = {}
@@ -200,11 +175,13 @@ def crop_image(request,*args,**kwargs):
             if cropY < 0:
                 cropY = 0
 
+
             crop_image = img[cropY:cropY + cropHeight,cropX:cropX + cropWidth]
 
             cv2.imwrite(url,crop_image)
 
-            user.profile_image.delete()
+            if user.profile_image != get_default_profile_image():
+                user.profile_image.delete()
             user.profile_image.save("profile_image.png",files.File(open(url,"rb")))
             user.save()
